Move ModelInstance class to VSUM project
