+++ bundles/framework/tools.vitruv.framework.vsum/src/tools/vitruv/framework/vsum/ModelRepository.xtend
@@ -1,6 +1,5 @@
 package tools.vitruv.framework.vsum
 
-import tools.vitruv.framework.util.datatypes.ModelInstance
 import tools.vitruv.framework.change.description.TransactionalChange
 import tools.vitruv.framework.propagation.ResourceAccess
 import tools.vitruv.framework.correspondence.CorrespondenceModel
