+++ tests/framework/tools.vitruv.framework.vsum.tests/src/tools/vitruv/framework/tests/vsum/CorrespondenceTest.xtend
@@ -10,7 +10,6 @@ import pcm_mockup.PInterface
 import pcm_mockup.Repository
 import tools.vitruv.framework.correspondence.Correspondence
 import tools.vitruv.framework.correspondence.CorrespondenceModel
-import tools.vitruv.framework.util.datatypes.ModelInstance
 import tools.vitruv.framework.vsum.InternalVirtualModel
 import uml_mockup.UInterface
 import uml_mockup.UPackage
@@ -209,7 +208,7 @@ class CorrespondenceTest {
 	}
 
 	def private <T extends EObject> T testLoadObject(InternalVirtualModel vsum, URI uri, Class<T> clazz) {
-		val ModelInstance instance = vsum.getModelInstance(uri)
+		val instance = vsum.getModelInstance(uri)
 		assertEquals(1, instance.resource.contents.size)
 		val root = instance.resource.contents.get(0)
 		assertThat(root, instanceOf(clazz))
